fix: 강의 찜 존재 여부 예외처리 추가

+++ src/main/java/com/reviewing/review/course/controller/CourseController.java
@@ -151,12 +151,18 @@ public class CourseController {
         }
 
         if (wished) { // wished=true -> 강의 찜 취소
+            if (courseService.checkCourseWishedByMember(courseId, memberId)) {
+                return ResponseEntity.status(605).body(null);
+            }
             courseService.removeCourseWish(courseId, memberId);
             CourseResponseDto courseResponseDto = courseService.findCourseById(courseId,memberId);
             return ResponseEntity.ok().body(courseResponseDto);
         }
 
         // wished=false -> 강의 찜
+        if (!courseService.checkCourseWishedByMember(courseId, memberId)) {
+            return ResponseEntity.status(606).body(null);
+        }
         courseService.createCourseWish(courseId, memberId);
         CourseResponseDto courseResponseDto = courseService.findCourseById(courseId,memberId);
         return ResponseEntity.ok().body(courseResponseDto);
